Add files via upload
actions.py
Added actions for fight, run and avoid. Will probably need to update fighting sequence to something more exciting. Now it is just choosing fight until you win or die.

Need to update show actions so that it only shows available actions and not all possible actions. Why show search corpse if no corpse?

MonsterRoom, LockedRoom and BossRoom.py
Added functionality for these rooms so that player encounters door and monster accordingly.

Bosskeyroom and Smallkeyroom.py
Added chest and corpse that can be searched.

monster.py
Added agility for monsters so that it can be used in fight for chance for monsters of avoiding damage